feat(core): add support for remote cache for garf reports
* Add redis cache support
* Add remote file support (s3, gs, etc) via smart_open

+++ libs/core/garf/core/cache.py
@@ -25,6 +25,8 @@ import pathlib
 import shutil
 from typing import Final
 
+import redis
+import smart_open
 from garf.core import exceptions, query_editor, report
 
 logger = logging.getLogger(__name__)
@@ -37,34 +39,33 @@ class GarfCacheFileNotFoundError(exceptions.GarfError):
 DEFAULT_CACHE_LOCATION: Final[str] = os.getenv(
   'GARF_CACHE_LOCATION', str(pathlib.Path.home() / '.garf/cache/')
 )
+DEFAULT_CACHE_TTL: Final[int] = os.getenv('GARF_CACHE_TTL_SECONDS', 3600)
 
 
 class GarfCache:
-  """Stores and loads reports from a cache instead of calling API.
-
-  Attribute:
-    location: Folder where cached results are stored.
-  """
+  """Stores and loads reports from a cache instead of calling API."""
 
   def __init__(
     self,
     location: str | None = None,
-    ttl_seconds: int = 3600,
+    ttl_seconds: int = DEFAULT_CACHE_TTL,
+    cache_provider=None,
   ) -> None:
-    """Stores and loads reports from a cache instead of calling API.
+    """Initializes new cache.
 
     Args:
-      location: Folder where cached results are stored.
+      location: Cache location.
       ttl_seconds: Maximum lifespan of cached objects.
+      cache_provider: Instantiated cache provider.
     """
-    self.location = pathlib.Path(location or DEFAULT_CACHE_LOCATION)
-    self.ttl_seconds = ttl_seconds
-
-  @property
-  def max_cache_timestamp(self) -> float:
-    return (
-      datetime.datetime.now() - datetime.timedelta(seconds=self.ttl_seconds)
-    ).timestamp()
+    self.location = location or DEFAULT_CACHE_LOCATION
+    self.ttl_seconds = int(ttl_seconds)
+    if cache_provider:
+      self.cache_provider = cache_provider
+    elif str(location).startswith('redis'):
+      self.cache_provider = RedisGarfCache(self.location, self.ttl_seconds)
+    else:
+      self.cache_provider = FileGarfCache(self.location, self.ttl_seconds)
 
   def load(
     self, query: query_editor.BaseQueryElements, args=None, kwargs=None
@@ -91,18 +92,7 @@ class GarfCache:
       else ''
     )
     hash_identifier = f'{query.hash}:{args_hash}:{kwargs_hash}'
-    cached_path = self.location / f'{hash_identifier}.json'
-    if (
-      cached_path.exists()
-      and cached_path.stat().st_ctime > self.max_cache_timestamp
-    ):
-      with open(cached_path, 'r', encoding='utf-8') as f:
-        data = json.load(f)
-      logger.debug('Report is loaded from cache: %s', str(cached_path))
-      loaded_report = report.GarfReport.from_json(json.dumps(data))
-      loaded_report.query_specification = query
-      return loaded_report
-    raise GarfCacheFileNotFoundError
+    return self.cache_provider.load(hash_identifier, query)
 
   def save(
     self,
@@ -119,7 +109,6 @@ class GarfCache:
       args: Query parameters.
       kwargs: Optional keyword arguments.
     """
-    self.location.mkdir(parents=True, exist_ok=True)
     args_hash = args.hash if args else ''
     kwargs_hash = (
       hashlib.md5(
@@ -129,14 +118,153 @@ class GarfCache:
       else ''
     )
     hash_identifier = f'{query.hash}:{args_hash}:{kwargs_hash}'
-    cached_path = self.location / f'{hash_identifier}.json'
-    logger.info('Report is saved to cache: %s', str(cached_path))
-    with open(cached_path, 'w', encoding='utf-8') as f:
+    result = self.cache_provider.save(fetched_report, hash_identifier)
+    logger.info('Report is saved to cache: %s', result)
+
+  @property
+  def size(self) -> int:
+    return 0
+
+
+class RedisGarfCache:
+  """Stores and loads reports from Redis."""
+
+  def __init__(
+    self,
+    location: str | None = None,
+    ttl_seconds: int = DEFAULT_CACHE_TTL,
+    redis_client: redis.Redis | None = None,
+  ) -> None:
+    self.location = location or DEFAULT_CACHE_LOCATION
+    self.r = redis_client or redis.Redis.from_url(
+      self.location, decode_responses=True
+    )
+    self.ttl_seconds = ttl_seconds
+
+  def load(self, hash_identifier: str, query) -> report.GarfReport:
+    """Loads report from cache based on query definition.
+
+    Args:
+      hash_identifier: Unique identifier of a query.
+
+    Returns:
+      Cached report.
+
+    Raises:
+      GarfCacheFileNotFoundError: If cached report not found
+    """
+    if not (data := self.r.get(hash_identifier)):
+      raise GarfCacheFileNotFoundError
+    logger.debug('Report is loaded from cache: %s', str(hash_identifier))
+    loaded_report = report.GarfReport.from_json(data)
+    loaded_report.query_specification = query
+    return loaded_report
+
+  def save(
+    self,
+    fetched_report: report.GarfReport,
+    hash_identifier: str,
+  ) -> None:
+    """Saves report to cache based on query definition.
+
+    Args:
+      fetched_report: Report to save.
+      hash_identifier: Unique identifier of a query.
+    """
+    data = json.dumps(fetched_report.to_list(row_type='dict'))
+    self.r.set(hash_identifier, data, ex=self.ttl_seconds)
+    logger.info('Report is saved to cache: %s', str(hash_identifier))
+    return hash_identifier
+
+
+class FileGarfCache:
+  """Stores and loads reports from remote or local file storage.
+
+  Attribute:
+    location: Folder where cached results are stored.
+  """
+
+  def __init__(
+    self,
+    location: str | None = None,
+    ttl_seconds: int = DEFAULT_CACHE_TTL,
+  ) -> None:
+    location = location or DEFAULT_CACHE_LOCATION
+    if '://' in str(location):
+      self.location = location
+      self.type = 'remote'
+    else:
+      self.location = pathlib.Path(location)
+      self.type = 'local'
+    self.ttl_seconds = ttl_seconds
+
+  @property
+  def max_cache_timestamp(self) -> float:
+    return (
+      datetime.datetime.now() - datetime.timedelta(seconds=self.ttl_seconds)
+    ).timestamp()
+
+  def load(self, hash_identifier: str, query) -> report.GarfReport:
+    """Loads report from cache based on query definition.
+
+    Args:
+      hash_identifier: Unique identifier of a query.
+
+    Returns:
+      Cached report.
+
+    Raises:
+      GarfCacheFileNotFoundError: If cached report not found
+    """
+    if self.type == 'local':
+      cached_path = self.location / f'{hash_identifier}.json'
+    else:
+      cached_path = f'{self.location}/{hash_identifier}.json'
+    if (
+      self.type == 'local'
+      and cached_path.exists()
+      and cached_path.stat().st_ctime > self.max_cache_timestamp
+    ):
+      with smart_open.open(cached_path, 'r', encoding='utf-8') as f:
+        data = json.load(f)
+    elif self.type == 'remote':
+      try:
+        with smart_open.open(cached_path, 'r', encoding='utf-8') as f:
+          data = json.load(f)
+      except Exception as e:
+        raise GarfCacheFileNotFoundError from e
+    else:
+      raise GarfCacheFileNotFoundError
+    logger.debug('Report is loaded from cache: %s', str(cached_path))
+    loaded_report = report.GarfReport.from_json(json.dumps(data))
+    loaded_report.query_specification = query
+    return loaded_report
+
+  def save(
+    self, fetched_report: report.GarfReport, hash_identifier: str
+  ) -> None:
+    """Saves report to cache based on query definition.
+
+    Args:
+      fetched_report: Report to save.
+      hash_identifier: Unique identifier of a query.
+    """
+    if self.type == 'local':
+      self.location.mkdir(parents=True, exist_ok=True)
+    if self.type == 'local':
+      cached_path = self.location / f'{hash_identifier}.json'
+    else:
+      cached_path = f'{self.location}/{hash_identifier}.json'
+    with smart_open.open(str(cached_path), 'w', encoding='utf-8') as f:
       json.dump(fetched_report.to_list(row_type='dict'), f)
+    logger.info('Report is saved to cache: %s', str(cached_path))
+    return str(cached_path)
 
   @property
   def size(self) -> int:
     total_size = 0
+    if self.type == 'remote':
+      return total_size
     for dirpath, dirnames, filenames in os.walk(self.location):
       for f in filenames:
         file_path = os.path.join(dirpath, f)
@@ -146,6 +274,8 @@ class GarfCache:
 
   def clean(self) -> None:
     """Removes all cached files."""
+    if self.type == 'remote':
+      return
     for item in self.location.iterdir():
       if item.is_dir():
         shutil.rmtree(item)
@@ -161,10 +291,12 @@ class GarfCache:
     Returns:
       Number of bytes pruned.
     """
+    pruned_file_size = 0
+    if self.type == 'remote':
+      return pruned_file_size
     prune_date = datetime.datetime.now(
       datetime.timezone.utc
     ) - datetime.timedelta(days=ttl)
-    pruned_file_size = 0
     for file_path in self.location.rglob('*.json'):
       if (
         file_path.is_file()
